exported calculation of data to another python script
